+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 - Added a `DBParser.onset_ranges` method that returns a list of tuples holding the onset ranges of all extracted fragments
 
 ### Fixed
-- Issue #54 (unreferenced variable); Issue #51 (updated databases); 
+- Issue #54 (unreferenced variable); Issue #51 (updated databases); Issue #58 (unfiltered fragment table); 
 
 ## [v0.6.3]
 ### Added